Fixed AllS4.R and tests for S4 stuff.
The setClass("S4Composition", representation(info="data.table")) was part of a test case to see if we can make an S4 class that has a data.table slot, then define methods that work on that data.table.

This was moved out of AllS4.R and into tests/tests.R. Things seem kosher now.



git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/datatable@206 edb9625f-4e0d-4859-8d74-9fd3b1da38cb