Briefly:
 * Support for "by" argument in merge.data.table (closes FR #1315)
 * Plumbing for using test_that

== merge.data.table ==

merge.data.table is acting more like merge.data.frame with the exception that
the default columns to merge on are the shared "key" columns between the two
data.tables that are being merged.

merge columns can now be specified using the "by" parameter.

if a user specified "by" parameter aren't already keys for the data.tables
(or are in the wrong order) the keys for x and y will be set to do a merge
on the "by" columns accordingly.

the "y" object can be a data.frame, which is coerced to a data.table if
necessary

== testing + test_that ==

An "all.equal.data.table" function was added in R/test.data.table.R
This lets us use test_that without defining our own expect_equivalent_dt
(or whatever) functions because the "all.equal.data.table" function will
be called when the first argument to the "normal" test_that::expect_equal
function is a data.table

I commented out my merge and S4 tests in inst/tests/tests.R since they are
now in inst/tests/test-merge.R and inst/tests/test-S4.R.

I also had to modify the functionality in inst/tests/test.R so that it
doesn't try to run every *.R file in inst/tests/tests.R because there was
an error thrown when test.data.table was trying to "drive" the test_that tests.

Sorry if we are stepping on each others feet, but I think we are planning to
transition to test_that (right?) -- so I reckon tests "borken out" to seperate
files will be the providence of test_that. Also, when that does happen
the lines that I have been commenting out should be eventually axed to
avoid bit rot.

Lastly, until Hadley pulls in my changes to test_that, you'll need to use
the version I've modified here for now so that a call like this will work:

expect_equal(a, b, info="some info here")

The current version doesn't take info or label arguments in the expect_equal
calls.

Here's my version (which is against Hadley's latest version on github)
https://github.com/lianos/test_that




git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/datatable@214 edb9625f-4e0d-4859-8d74-9fd3b1da38cb