Tweaking support for silencing dt[, some := thing] printing via a "history-hack".
There is a check .onLoad to see if history() can be run successfully which determines
if this hack should be enabled or not -- I think it will always work unless R is
run with readline support (sadly, ESS disables readline support, but c'est la vie).

Matthew: Let me know if you like the way this is going, I can take a crack at updating
relevant docs/FAQs if you want to stick with this, otherwise we can back this out.



git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/datatable@643 edb9625f-4e0d-4859-8d74-9fd3b1da38cb